Backend

+++ backend/src/Controller/CredentialsController.php
@@ -21,16 +21,39 @@ class CredentialsController extends AbstractController
         $this->crudService = $crudService;
     }
 
-    public function index(): JsonResponse
+    public function index(Request $request, EntityManagerInterface $em): JsonResponse
     {
-        return $this->crudService->index(Credentials::class);
-    }
+        $data = $request->getContent();
+        $dataJson = json_decode($data, true);
 
-//    public function show(Request $request): JsonResponse
-//    {
-//        $data = $request->getContent();
-//        return $this->crudService->show(Credentials::class, $data);
-//    }
+        if (!isset($dataJson['email'])) {
+            return new JsonResponse(['message' => 'Email is required', 'code' => 400], Response::HTTP_BAD_REQUEST);
+        }
+
+        $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
+        if (!$user) {
+            return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $credentialsForUser = $em->getRepository(Credentials::class)->findBy(['user' => $user]);
+
+        $jsonEntities = [];
+
+        foreach ($credentialsForUser as $credential) {
+            $jsonEntities[] = [
+                'id' => $credential->getId(),
+                'title' => $credential->getTitle(),
+                'url' => $credential->getUrl(),
+                'username' => $credential->getUsername(),
+                'notes' => $credential->getNotes(),
+                'category' => $credential->getCategory()?->getTitle(),
+                'password' => $credential->getPassword(),
+                'user' => $credential->getUser()->getEmail()
+            ];
+        }
+
+        return new JsonResponse(['message' => $jsonEntities, 'code' => 200], Response::HTTP_OK);
+    }
 
     public function create(Request $request, EntityManagerInterface $em): JsonResponse
     {
@@ -51,14 +74,14 @@ class CredentialsController extends AbstractController
             return new JsonResponse(['message' => 'Password is required', 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
-        $entity = $em->getRepository(Credentials::class)->findOneBy(['title' => $dataJson['title']]);
-        if ($entity) {
-            return new JsonResponse(['message' => "Credential with title {$dataJson['title']} already exists.", 'code' => 400], Response::HTTP_BAD_REQUEST);
-        }
-
         $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
         if (!$user) {
             return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $entity = $em->getRepository(Credentials::class)->findOneBy(['title' => $dataJson['title'], 'user' => $user]);
+        if ($entity) {
+            return new JsonResponse(['message' => "Credential with title '[{$dataJson['title']}' for user {$user} already exists.", 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
         if (isset($dataJson['category'])) {
@@ -83,14 +106,14 @@ class CredentialsController extends AbstractController
             return new JsonResponse(['message' => 'Title is required', 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
-        $entity = $em->getRepository(Credentials::class)->findOneBy(['title' => $dataJson['title']]);
-        if (!$entity) {
-            return new JsonResponse(['message' => "Credential with title {$dataJson['title']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
-        }
-
         $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
         if (!$user) {
             return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $entity = $em->getRepository(Credentials::class)->findOneBy(['title' => $dataJson['title'],'user' => $user]);
+        if (!$entity) {
+            return new JsonResponse(['message' => "Credential with title {$dataJson['title']} for user $user not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
         }
 
         return $this->crudService->update(Credentials::class, $data, $entity);
@@ -101,11 +124,23 @@ class CredentialsController extends AbstractController
         $data = $request->getContent();
         $dataJson = json_decode($data, true);
 
+        if (!isset($dataJson['email'])) {
+            return new JsonResponse(['message' => 'Email is required', 'code' => 400], Response::HTTP_BAD_REQUEST);
+        }
+
         if (!isset($dataJson['title'])) {
             return new JsonResponse(['message' => 'Title is required', 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
-        $entity = $em->getRepository(Credentials::class)->findOneBy(['title' => $dataJson['title']]);
+        $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
+        if (!$user) {
+            return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $entity = $em->getRepository(Credentials::class)->findOneBy(['title' => $dataJson['title'], 'user' => $user]);
+        if (!$entity) {
+            return new JsonResponse(['message' => "Credential with title {$dataJson['title']} for user $user not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
 
         return $this->crudService->delete(Credentials::class, $entity);
     }
