+++ backend/src/Controller/GroupController.php
@@ -50,11 +50,6 @@ class GroupController extends AbstractController
         return new JsonResponse(['message' => $jsonEntities, 'code' => 200], Response::HTTP_OK);
     }
 
-//    public function show(int $id): JsonResponse
-//    {
-//        return $this->crudService->show(Group::class, $id);
-//    }
-
     public function create(Request $request, EntityManagerInterface $em): JsonResponse
     {
         $data = $request->getContent();
@@ -67,14 +62,14 @@ class GroupController extends AbstractController
             return new JsonResponse(['message' => 'Email is required.', 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
-        $entity = $em->getRepository(Group::class)->findOneBy(['title' => $dataJson['title']]);
-        if ($entity) {
-            return new JsonResponse(['message' => "Group with title {$dataJson['title']} already exists.", 'code' => 400], Response::HTTP_BAD_REQUEST);
-        }
-
         $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
         if (!$user) {
             return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $entity = $em->getRepository(Group::class)->findOneBy(['title' => $dataJson['title'], 'user' => $user]);
+        if ($entity) {
+            return new JsonResponse(['message' => "Group with title '{$dataJson['title']}' for user {$user} already exists.", 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
         return $this->crudService->create(Group::class, $data, $user);
@@ -89,24 +84,45 @@ class GroupController extends AbstractController
             return new JsonResponse(['message' => 'Title is required', 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
-        $entity = $em->getRepository(Group::class)->find($id);
-        if (!$entity) {
-            return new JsonResponse(['message' => "Group with id $id not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        if (!isset($dataJson['email'])) {
+            return new JsonResponse(['message' => 'Email is required.', 'code' => 400], Response::HTTP_BAD_REQUEST);
         }
 
-        $other = $em->getRepository(Group::class)->findOneBy(['title' => $dataJson['title']]);
+        $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
+        if (!$user) {
+            return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $entity = $em->getRepository(Group::class)->findOneBy(['id' => $id, 'user' => $user]);
+        if (!$entity) {
+            return new JsonResponse(['message' => "Group with id $id for user $user not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $other = $em->getRepository(Group::class)->findOneBy(['title' => $dataJson['title'], 'user' => $user]);
         if ($other) {
-            return new JsonResponse(['message' => "Group with title {$dataJson['title']} already exists.", 'code' => 400], Response::HTTP_NOT_FOUND);
+            return new JsonResponse(['message' => "Group with title {$dataJson['title']} for user $user already exists.", 'code' => 400], Response::HTTP_NOT_FOUND);
         }
 
         return $this->crudService->update(Group::class, $data, $entity);
     }
 
-    public function delete(int $id, EntityManagerInterface $em): JsonResponse
+    public function delete(int $id, Request $request, EntityManagerInterface $em): JsonResponse
     {
-        $entity = $em->getRepository(Group::class)->find($id);
+        $data = $request->getContent();
+        $dataJson = json_decode($data, true);
+
+        if (!isset($dataJson['email'])) {
+            return new JsonResponse(['message' => 'Email is required.', 'code' => 400], Response::HTTP_BAD_REQUEST);
+        }
+
+        $user = $em->getRepository(User::class)->findOneBy(['email' => $dataJson['email']]);
+        if (!$user) {
+            return new JsonResponse(['message' => "User with mail {$dataJson['email']} not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+        }
+
+        $entity = $em->getRepository(Group::class)->findOneBy(['id' => $id, 'user' => $user]);
         if (!$entity) {
-            return new JsonResponse(['message' => "Group with id $id not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
+            return new JsonResponse(['message' => "Group with id $id for user $user not found.", 'code' => 404], Response::HTTP_NOT_FOUND);
         }
 
         return $this->crudService->delete(Group::class, $entity);
